Move sync command integration tests to a dedicated file
Previously, the tool's command line interface was only ever intended to provide the "sync" functionality so a single
"all" integration test file was reasonable. Now that additional distinct command line functionality is planned, it will
be better to separate the integration tests into multiple files to make it easier for the developer to navigate the code.
