+++ .prettierignore
@@ -1,3 +1,3 @@
 # See: https://prettier.io/docs/en/ignore.html#ignoring-files-prettierignore
 
-/test/testdata/test_all/golden/logs/
+/test/testdata/test_sync/golden/logs/
