+++ test/testdata/test_sync/golden/library_index.json
