+++ test/testdata/test_sync/golden/logs/generate/github.com/arduino-libraries/SpacebrewYun/index.html
