+++ test/testdata/test_sync/golden/logs/update/github.com/arduino-libraries/ArduinoCloudThing/index.html
